fix: Refactor fixed Switch into own component for reuse
