+++ src/utils/use-gate.ts
@@ -1,0 +1,12 @@
+import {useState, useEffect} from 'react';
+
+/** Returns false at first, then true after a set time */
+export default function useGate(delayMs: number): boolean {
+  const [gate, setGate] = useState(false);
+  useEffect(() => {
+    setTimeout(() => {
+      setGate(true);
+    }, delayMs);
+  });
+  return gate;
+}
